Fix "last submission time" parsing issue
We are affected by the following python-dateutil issue:
https://github.com/dateutil/dateutil/issues/163

The symptoms are described here:
https://openedx.atlassian.net/jira/software/projects/BTR/boards/641?selectedIssue=BTR-25

In a nutshell: when creating an exercise with a 10 seconds answer delay
on a server that is configured with a non-UTC tzdata (but UTC TIME_ZONE
django setting), students are faced with error messages stating that
they have to wait 4h 59min 50s.

The deeper issue is that dateutil incorrectly parses the datetime
object. This issue is resolved by simply upgrading python-dateutil.

I realise that dateutil was constrained to an older version before. We
will have to see whether this change causes production errors.